meritpath-backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3
python-dotenv>=1.0.0
supabase>=1.0.3
httpx>=0.24.1
boto3>=1.28.0
python-jose[cryptography]>=3.3.0
pydantic>=1.8.2
pyjwt>=2.3.0